+++ airflow/cli/commands/task_command.py
@@ -259,7 +259,7 @@ def _run_task_by_executor(args, dag: DAG, ti: TaskInstance) -> None:
     executor.end()
 
 
-def _run_task_by_local_task_job(args, ti: TaskInstance) -> TaskReturnCode | None:
+def _run_task_by_local_task_job(args, ti: TaskInstance | TaskInstancePydantic) -> TaskReturnCode | None:
     """Run LocalTaskJob, which monitors the raw task execution process."""
     job_runner = LocalTaskJobRunner(
         job=Job(dag_id=ti.dag_id),