+++ airflow/contrib/operators/__init__.py
@@ -1049,15 +1049,6 @@ __deprecated_classes = {
             "airflow.providers.google.cloud.operators.pubsub.PubSubDeleteTopicOperator"
         ),
     },
-    "qubole_check_operator": {
-        "QuboleCheckOperator": "airflow.providers.qubole.operators.qubole_check.QuboleCheckOperator",
-        "QuboleValueCheckOperator": (
-            "airflow.providers.qubole.operators.qubole_check.QuboleValueCheckOperator"
-        ),
-    },
-    "qubole_operator": {
-        "QuboleOperator": "airflow.providers.qubole.operators.qubole.QuboleOperator",
-    },
     "redis_publish_operator": {
         "RedisPublishOperator": "airflow.providers.redis.operators.redis_publish.RedisPublishOperator",
     },